+++ library/src/schemas/blob/blob.ts
@@ -1,6 +1,10 @@
 import { ValiError } from '../../error/index.ts';
 import type { BaseSchema, Pipe } from '../../types.ts';
-import { executePipe, getErrorAndPipe } from '../../utils/index.ts';
+import {
+  executePipe,
+  getErrorAndPipe,
+  getPipeInfo,
+} from '../../utils/index.ts';
 
 /**
  * Blob schema type.
@@ -71,7 +75,7 @@ export function blob(
       }
 
       // Execute pipe and return output
-      return executePipe(input, pipe, { ...info, reason: 'blob' });
+      return executePipe(input, pipe, getPipeInfo(info, 'blob'));
     },
   };
 }