+++ library/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './executePipe/index.ts';
 export * from './getCurrentPath/index.ts';
 export * from './getErrorAndPipe/index.ts';
+export * from './getPipeInfo/index.ts';
 export * from './isLuhnAlgo/index.ts';